Refactor query handling onto the base query since enough is the same
