--- xpublish_edr/cube/params.py
@@ -1,40 +1,0 @@
-"""Query parameter model for EDR cube queries"""
-
-import xarray as xr
-from pydantic import Field, field_validator
-
-from xpublish_edr.base_query import BaseEDRQuery
-from xpublish_edr.format import cube_formats
-from xpublish_edr.geometry.common import project_bbox
-
-
-class EDRCubeQuery(BaseEDRQuery):
-    """
-    Capture query parameters for EDR cube queries
-    """
-
-    bbox: tuple[float, float, float, float] = Field(
-        ...,
-        title="Bounding box in minx, miny, maxx, maxy",
-        description="Bounding box for the query",
-    )
-
-    @field_validator("format", mode="before")
-    def validate_format(cls, v):
-        """Validate the format is a valid cube format"""
-        if v not in cube_formats().keys():
-            raise ValueError(f"Invalid format: {v}")
-        return v
-
-    @field_validator("bbox", mode="before")
-    def validate_bbox(cls, v):
-        """Validate the bbox is a tuple of 4 floats"""
-        if isinstance(v, str):
-            return tuple(float(v.strip()) for v in v.split(","))
-        if isinstance(v, list):
-            return tuple(float(v.strip()) for v in v[0].split(","))
-        raise ValueError(f"Invalid bbox: {v}")
-
-    def project_bbox(self, ds: xr.Dataset) -> tuple[float, float, float, float]:
-        """Project the bbox to the dataset's CRS"""
-        return project_bbox(ds, self.crs, self.bbox)
